rRefactor AuthContext: separate context and provider, cleaning up unused imports, fix ESLint and fast refresh
