+++ src/components/ProtectedRoute.jsx
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 import { Navigate } from "react-router-dom";
-import { AuthContext } from "../context/AuthContext";
+import { AuthContext } from "../context/AuthProvider";
 
 export function ProtectedRoute({ children }) {
   const { user } = useContext(AuthContext);
