+++ src/context/AuthContext.js
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+
+export const AuthContext = createContext(null);
+
