+++ src/pages/BookingList.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../context/AuthProvider";
 import { getUserBookings, deleteBooking } from "../services/api";
 
 function BookingList() {
