+++ src/pages/Bookings.jsx
@@ -1,4 +1,4 @@
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../context/AuthProvider";
 import { apiGet } from "../components/api";
 import { useState, useEffect } from "react";
