+++ src/pages/Dashboard.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from "react";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../context/AuthProvider";
 import { Link } from "react-router-dom";
 import { apiGet } from "../components/api";
