+++ src/pages/Profile.jsx
@@ -1,4 +1,4 @@
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../context/AuthProvider";
 import { useState } from "react";
 import { updateProfile } from "../components/api";
