+++ src/pages/UserProfile.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../context/AuthProvider";
 import { updateUser } from "../services/api";
 
 function UserProfile() {
